# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sadhak API",
    "version": "dev",
    "description": "Sadhak keeps a dependency graph of your systems and gates the changes\nthat would break them.\n\n**Verdicts are deterministic.** The same change against the same graph\nreturns the same answer, every time, with no model in the path. An edge\na model inferred can raise a warning and can never produce a BLOCK.\n\n**Tenancy is resolved from the credential, never from input.** Every\norg-scoped route also answers at `/api/orgs/{orgId}/…`, where `{orgId}`\nis *asserted* against the org your credential resolved to — a mismatch\nis a **404, not a 403**, so a wrong org id is indistinguishable from one\nthat does not exist."
  },
  "servers": [
    {
      "url": "https://api.sadhak.online",
      "description": "Production"
    }
  ],
...
  "components": {
    "schemas": {
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details. Served as application/problem+json.",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "instance",
          "requestId"
        ]
      },
...
      "SignUp": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "password": {
            "type": "string"
          },
          "company": {
            "type": "string",
            "maxLength": 120
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ],
        "additionalProperties": false
      },
      "SignIn": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "An org-scoped key, `sadhak_…`. Also accepted as `Authorization: Bearer`. A key can never hold more capability than the person who created it."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sadhak_session",
        "description": "What the web app uses. Set by POST /api/auth/signin."
      }
    }
  },
  "security": [
    {
      "apiKey": []
    },
    {
      "session": []
    }
  ],
  "paths": {
    "/api/auth/signup": {
      "post": {
        "summary": "Create an account and its first organisation",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUp"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created. Sets the session cookie."
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "That email already has an account.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Carries Retry-After.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/signin": {
      "post": {
        "summary": "Start a session",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in. Sets the session cookie."
          },
          "401": {
            "description": "Wrong email or password — the two are not distinguished.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Ten a minute, because that is where guessing a password becomes worth someone's time.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "The current user, their orgs, and their capabilities",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "The session's identity and what it may do."
          },
          "401": {
            "description": "No session.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  }
}